requirements.txt
#put all requirements here

amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
boto3==1.2.1
botocore==1.3.5
celery==3.1.19
defusedxml==0.4.1
Django==1.9
django-allauth==0.24.1
django-rest-auth==0.6.0
djangorestframework==3.3.1
docutils==0.12
jmespath==0.9.0
kombu==3.0.29
oauthlib==1.0.3
Pillow==3.0.0
python-dateutil==2.4.2
python3-openid==3.0.9
pytz==2015.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
django-extensions==1.5.9
Werkzeug==0.11.2
drf-nested-decorator==0.3
opbeat==3.1.3
django-geoposition==0.2.2
django-rest-swagger
sendgrid-django
django-filter
django-crispy-forms